Moved the Loader and HeaderMessage components to the common module.
