+++ src/app/app.module.ts
@@ -49,11 +49,9 @@ import { ErrorPage } from "./components/errorPage/errorPage.component";
 import { NewItem } from "./components/newItem/newItem.component";
 import { MyPosts } from "./components/myPosts/myPosts.component";
 import { FullList } from "./components/fullList/fullList.component";
-import { Loader } from "./components/loader/loader.component";
 import { AboutApp } from "./components/aboutApp/aboutApp.component";
 import { SearchResults } from "./components/searchResults/searchResults.component";
 import { AdminDashboard } from "./components/adminDashboard/adminDashboard.component";
-import { HeaderMessage } from "./components/headerMessage/headerMessage.component";
 import { NotificationsTab } from "./components/notifications/notifications.component";
 import { SettingsPage } from "./components/settings/settings.component";
 import { SiteMap } from "./components/siteMap/siteMap.component";
@@ -85,11 +83,9 @@ import { LoginPage } from "./components/loginPage/loginPage.component";
     NewItem,
     MyPosts,
     FullList,
-    Loader,
     AboutApp,
     SearchResults,
     AdminDashboard,
-    HeaderMessage,
     NotificationsTab,
     SettingsPage,
     SiteMap,
