+++ src/app/common/components/headerMessage/headerMessage.component.spec.ts
@@ -43,7 +43,6 @@ import { ServiceWorkerModule } from "@angular/service-worker";
 import { FontAwesomeModule } from "@fortawesome/angular-fontawesome";
 
 import { HeaderMessage } from "./headerMessage.component";
-import { AppCommonModule } from "@app/common/common.module";
 
 describe("HeaderMessage", () => {
   // Before each test, configure testing environment
@@ -57,7 +56,6 @@ describe("HeaderMessage", () => {
         HttpClientModule,
         ServiceWorkerModule.register("sw.js", { enabled: false }),
         FontAwesomeModule,
-        AppCommonModule,
       ],
       declarations: [HeaderMessage],
       providers: [{ provide: APP_BASE_HREF, useValue: "/" }],
