+++ src/app/components/adminFilters/adminFilters.component.spec.ts
@@ -48,7 +48,7 @@ import { of } from "rxjs";
 import { AdminFilters } from "./adminFilters.component";
 import { PopUp } from "@common/components/popUp/popUp.component";
 import { AuthService } from "../../common/services/auth.service";
-import { Loader } from "../loader/loader.component";
+import { Loader } from "../../common/components/loader/loader.component";
 import { mockAuthedUser } from "@tests/mockData";
 import { ApiClientService } from "@common/services/apiClient.service";
 import { AppCommonModule } from "@app/common/common.module";
