+++ src/app/components/adminReports/adminReports.component.spec.ts
@@ -47,7 +47,7 @@ import { By } from "@angular/platform-browser";
 
 import { AdminReports } from "./adminReports.component";
 import { AuthService } from "../../common/services/auth.service";
-import { Loader } from "../loader/loader.component";
+import { Loader } from "../../common/components/loader/loader.component";
 import { mockAuthedUser } from "@tests/mockData";
 import { Report } from "@app/interfaces/report.interface";
 import { ApiClientService } from "@common/services/apiClient.service";
