+++ src/app/components/fullList/fullList.component.spec.ts
@@ -46,7 +46,7 @@ import { ActivatedRoute, Router, RouterModule, UrlSegment } from "@angular/route
 import { FullList } from "./fullList.component";
 import { PopUp } from "@common/components/popUp/popUp.component";
 import { SinglePost } from "@common/components/post/post.component";
-import { Loader } from "../loader/loader.component";
+import { Loader } from "../../common/components/loader/loader.component";
 import { ApiClientService } from "../../common/services/apiClient.service";
 import { SWManager } from "../../common/services/sWManager.service";
 import { AppCommonModule } from "@app/common/common.module";
