+++ src/app/components/mainPage/mainPage.component.spec.ts
@@ -45,7 +45,7 @@ import { of } from "rxjs";
 
 // App imports
 import { MainPage } from "./mainPage.component";
-import { Loader } from "../loader/loader.component";
+import { Loader } from "../../common/components/loader/loader.component";
 import { PopUp } from "@common/components/popUp/popUp.component";
 import { SinglePost } from "@common/components/post/post.component";
 import { ApiClientService } from "../../common/services/apiClient.service";
